requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pyTelegramBotAPI>=4.0.0
requests>=2.25.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
cryptography>=3.4.8
mysql-connector-python>=8.0.0